Improvement of the readability of functions
Added more comments, changed some variables names on allowedCells and
added more test.

*PROBLEMS*

The results of the two latest tests doesn't match with the results
that have been given in the documentation of this challenge.

I think that is a problem related with the indexes of the board and snake.
The last test "test5", is a modification of Test 2, where I added one column
to the board, that modification allowed to achieve the desired result.